:art: Style : converter 분리
<footer>
- 관련: #294
